--- a/routing/matcher.go
+++ b/routing/matcher.go
@@ -28,7 +28,7 @@ type leafRequestMatcher struct {
 	path string
 }
 
-func (m *leafRequestMatcher) IsMatch(value interface{}) bool {
+func (m *leafRequestMatcher) Match(value interface{}) bool {
 	v := value.(*pathMatcher)
 	l := matchLeaves(v.leaves, m.r, m.path)
 
@@ -248,13 +248,7 @@ func newMatcher(rs []*Route, o MatchingOptions) (*matcher, []*definitionError) {
 
 // matches a path in the path trie structure.
 func matchPathTree(tree *pathmux.Tree, path string, lrm *leafRequestMatcher) (leafMatchers, map[string]string) {
-	var matcher pathmux.Matcher = nil
-
-	if (lrm != nil) {
-		matcher = lrm
-	}
-
-	v, params := tree.Lookup(path, matcher)
+	v, params := tree.LookupMatcher(path, lrm)
 	if v == nil {
 		return nil, nil
 	}