--- a/routing/matcher_test.go
+++ b/routing/matcher_test.go
@@ -653,7 +653,8 @@ func TestMatchPathTreeNoMatch(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	m, p := matchPathTree(tree, "/some/wrong/path", nil)
+
+	m, p := matchPathTree(tree, "/some/wrong/path", &leafRequestMatcher{})
 	if len(m) != 0 || len(p) != 0 {
 		t.Error("failed not to match path")
 	}
@@ -671,7 +672,7 @@ func TestMatchPathTree(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	m, p := matchPathTree(tree, "/some/path", nil)
+	m, p := matchPathTree(tree, "/some/path", &leafRequestMatcher{&http.Request{}, ""})
 	if len(m) != 1 || len(p) != 0 || m[0] != pm0.leaves[0] {
 		t.Error("failed to match path", len(m), len(p))
 	}
@@ -689,7 +690,7 @@ func TestMatchPathTreeWithWildcards(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	m, p := matchPathTree(tree, "/some/path/and/params", nil)
+	m, p := matchPathTree(tree, "/some/path/and/params", &leafRequestMatcher{&http.Request{}, ""})
 	if len(m) != 1 || len(p) != 2 || m[0] != pm0.leaves[0] ||
 		p["param0"] != "and" || p["param1"] != "params" {
 		t.Error("failed to match path", len(m), len(p))