fix #116 with use changes in https://github.com/zalando/pathmux/pull/2