@ewgRa 2016-04-18
1. Changes done. Can you benchmark skipper/routing?
For BenchmarkGeneric I have +11% with this final changes.

2. My laptop can done only BenchmarkGeneric, but on BenchmarkPathTree1 get kill signal and "out of memory" in syslog. Is this expected and it is only GC problem or there memory leaks?

3. Also see pathmux issue comments.