@ewgRa 2016-04-19
So, I just reduce random routes counts in tests and run some benchmarks.

```
./benchcmp master-bench.txt fix-matcher-matchleaves-twice-call-bench.txt
benchmark                          old ns/op     new ns/op     delta
BenchmarkGeneric-4                 160205        193990        +21.09%
BenchmarkPathTree1-4               1085          1590          +46.54%
BenchmarkPathTree2-4               1367          1866          +36.50%
BenchmarkPathTree3-4               2198          3102          +41.13%
BenchmarkPathTree4-4               2754          3906          +41.83%
BenchmarkConstructionGeneric-4     208088        288422        +38.61%
BenchmarkConstructionMass-4        133039264     181532283     +36.45%
```

```
./benchcmp master-bench.txt fix-matcher-bench.txt 
benchmark                          old ns/op     new ns/op     delta
BenchmarkGeneric-4                 160205        192500        +20.16%
BenchmarkPathTree1-4               1085          1544          +42.30%
BenchmarkPathTree2-4               1367          1811          +32.48%
BenchmarkPathTree3-4               2198          3155          +43.54%
BenchmarkPathTree4-4               2754          3438          +24.84%
BenchmarkConstructionGeneric-4     208088        292211        +40.43%
BenchmarkConstructionMass-4        133039264     184731451     +38.85%
```

I interpret it as there is no big difference if call matchLeaves twice, or try to save this call by introduce return match value in Matcher.

But this "save call" way have twice degradation in pathmux performance if there is good benchmarks there.

Also as total skipper performance degradation about ~30%, that seems huge for me.

So, my opinion, revert last commit that save this matchLeaves call by return value from Matcher in skipper and pathmux, and just add missed call Matcher for catchAll case in pathmux.